Refactor artwork upload
