+++ apps/web/src/components/Artwork/index.ts
@@ -1,0 +1,2 @@
+export * from './ArtworkPreview'
+export * from './ArtworkUpload'
