+++ apps/web/src/hooks/index.ts
@@ -5,3 +5,5 @@ export * from './useIsMounted'
 export * from './usePrevious'
 export * from './useTimeout'
 export * from './useVotes'
+export * from './useArtworkPreview'
+export * from './useArtworkUpload'
